Module google_pagespeedonline5::api

Structs

AuditRefs
A light reference to an audit by id, used to group and weight audits in a given category.

Bucket
A proportion of data in the total distribution, bucketed by a min/max percentage. Each bucket's range is bounded by min <= x < max, In millisecond.

Categories
The categories in a Lighthouse run.

CategoryGroupV5
Message containing a category

ConfigSettings
Message containing the configuration settings for the Lighthouse run.

Environment
Message containing environment configuration for a Lighthouse run.

I18n
Message containing the i18n data for the LHR - Version 1.

LighthouseAuditResultV5
An audit's result object in a Lighthouse result.

LighthouseCategoryV5
A Lighthouse category.

LighthouseResultV5
The Lighthouse result object.

PagespeedApiLoadingExperienceV5
The CrUX loading experience object that contains CrUX data breakdowns.

PagespeedApiPagespeedResponseV5
The Pagespeed API response object.

PagespeedInsights
Central instance to access all PagespeedInsights related resource activities

PagespeedVersion
The Pagespeed Version object.

PagespeedapiMethods
A builder providing access to all methods supported on pagespeedapi resources. It is not used directly, but through the PagespeedInsights hub.

PagespeedapiRunpagespeedCall
Runs PageSpeed analysis on the page at the specified URL, and returns PageSpeed scores, a list of suggestions to make that page faster, and other information.

RendererFormattedStrings
Message holding the formatted strings used in the renderer.

RuntimeError
Message containing a runtime error config.

StackPack
Message containing Stack Pack information.

Timing
Message containing the performance timing data for the Lighthouse run.

UserPageLoadMetricV5
A CrUX metric object for a single metric and form factor.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.